requirements.txt
streamlit>=1.35.0
langchain>=0.2.0
langchain-community>=0.2.0
langchain-huggingface>=0.0.3
langchain-groq>=0.1.6
faiss-cpu>=1.8.0
sentence-transformers>=3.0.0
pypdf>=4.2.0
python-dotenv>=1.0.0
